+++ components/vf-search-client-side/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
-# 1.0.0-alpha.2
+## 1.0.0-alpha.3
+
+* adds support for `autofocus` on search input
+
+## 1.0.0-alpha.2
 
 * Improve documentation
 
-# 1.0.0-alpha.1
+## 1.0.0-alpha.1
 
 * Initial prototype