+++ components/vf-section-header/CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.3.0
+
+- updates spacing design tokens
+- requires `v2.0.0` of the `vf-design-tokens` package or newer.
+
+### 1.2.2
+
+- adds `id` for anchor support
+
 ### 1.2.1
 
 - makes the `--is-link` variant use CSS grid so the arrow SVG icon is always to the top, right of the text.