+++ components/vf-utility-classes/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2.0.0
+
+- all utility classes that make use of spacing tokens now use numerical naming.
+- updates documentation pages, as needed.
+- requires `v2.0.0` of the `vf-design-tokens` package or newer.
+
 ### 1.1.0
 
 - adds utility classes for grid row and columns gaps