+++ tools/vf-sass-compilation/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.2
+
+- adds ESLint task and config
+  - https://github.com/visual-framework/vf-core/issues/1143
+
 ### 1.0.1
 
 - dependency bump